[BUGFIX] iwlwifi: mvm: fix "failed to remove key" message
When the GTK is installed, we install it to HW with the
station ID of the AP.

Mac80211 will try to remove it only after the AP sta is
removed, which will result in a failure to remove key
since we do not have any station for it.

This is a valid situation, but a previous commit removed
the early return and added a return with error value, which
resulted in an error message that is confusing to users.

Remove the error return value.

type=bugfix
fixes=I5ded4acc6ee158e7b93d55a231dcac81dde6cbd3
ticket=kernel.org:198357

Change-Id: Ide9332ac7960ac5dd72a4fdd6e1ff855fc6ea3a9
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/156561
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/158314
x-iwlwifi-stack-dev: 6ec02fd2d3920b625675c929e49af1d3f6b5f461