[BUGFIX] iwlwifi: mvm: Move unused phy's to a default channel
When immediate quiet bit is set in CSA, the entire channel is blocked
by the firmware. It is expected that all the MACs will evacuate the
channel and the phy will be eventually either moved or removed.
Currently, the phy context is just unreferenced and thus, the quiet bit
is kept set and it will be impossible to TX on this phy, if we will need
to reuse it in the future.
This can be seen when doing a channel switch with mode=1 (quiet) twice
from channel X to Y and then back to channel X.
Fix that, by moving the phy context to a default channel when not
referenced anymore.

type=bugfix
fixes=unknown
ticket=none

Change-Id: I7745f9264c487ff82d67d6e907d0636ff4ff8891
Signed-off-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/164623
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: d01bcd8ca670fbd788160c64fc384097465f0a70