[BUGFIX] mac80211: Fix setting of SAE authentication timeout
In commit 69cc9c8821a9 ("mac80211: Adjust SAE authentication timeout")
the SAE timeout was adjusted to adhere with the new specification
definitions. However, this did not consider the case that the driver
does not support Tx status report, and did not consider the fact that
in case that Tx status report is indeed supported, the timeout is reset
to a lower value.

Fix these issues by only setting the SAE timeout for the case that Tx
status is not supported, and in addition, in case that the Tx status
is supported, set the SAE timeout only after the frame is ACKed.

type=bugfix
ticket=none
fixes=I19930b764dc78ab837ca99ee38baa70ed21c1651

Change-Id: Ib0b1185b87bf954c60e7883103bad649eea0fd30
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 162a6952f94c68ab42830430c7f58070c5490e38