[BUGFIX] iwlwifi: mvm: set wep key for all stations in soft ap mode
When operating as a soft ap with wep security, the key was not
configured to the fw for the stations, based on the fact that the
key will be specified in the tx command.

However, in the new tx api the tx command does not include the key,
which resulted in all data frames going out un-encrypted.

Fix it by configuring the key for all the stations as they are added.

type=bugfix
ticket=jira:WIFI-9820
fixes=unknown

Change-Id: I204f5be92854e9637ea3774610e25dcdf3bc046e
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/176525
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/177743
x-iwlwifi-stack-dev: b3e9ceda0ac815869b8e92387088bed18ab81d24