[BUGFIX] iwlwifi: pcie: avoid empty free RB queue
If all free RB queues are empty, the Driver
will never restock the free RB queue.
That's because the restocking happens in the Rx flow,
and if the free queue is empty there will be no Rx.

Although there's a background worker (a.k.a allocator)
allocating memory for RBs so that the Rx handler
can restock them, the worker may run only after
the free queue has become empty (and then
it is too late for restocking as explained above)

There is a solution for that called 'emergency':
If the number of used RB's reaches half the amount of all
RB's, the Rx handler will not wait for the allocator
but immediately allocate memory for the used RB's and
restock the free queue.

But, since the used RB's is per queue, it may happen
that the used RB's are spread between the queues such
that the emergency check will fail for each of the queues
(and still run out of RBs, causing the above symptom).

To fix it, move to emergency mode if the sum of *all*
used RBs (for all Rx queues) reaches half the amount
of all RB's

type=bugfix
fixes=unknown
ticket=jira:WIFI-9553

Change-Id: I711058183ae488863cab6d6df172142d2c805594
Signed-off-by: Shaul Triebitz <shaul.triebitz@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/176848
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/178457
x-iwlwifi-stack-dev: 3777bb3398390086c5794928febf063124f296bb