nl80211: NDP: clarify message handling in nl80211_nan_ndp()
Only in the case of NL80211_NAN_NDP_OPER_REQ do we send a
response message, but if that ever changes the condition
would have to be updated. Change the condition to instead
be on whether or not a message should be sent, to clarify
the code.

type=cleanup
ticket=jira:WLWG-7875

Change-Id: Ic248619ef0a73be92241cd27514b38932fb88029
Signed-off-by: Johannes Berg <johannes.berg@intel.com>x-iwlwifi-stack-dev: f584bcb80628fc97d9b5efdfee8f7e8cab6d500a