[BUGFIX] iwlwifi: mvm: don't use iwl_geo_tx_power_profiles_cmd_v1
We don't need to allocate a new structure and copy the data from the
new one into it, because they are the same, except that the new
version has one more element at the end.  We can simply send the same
structure but pass the size of the old one when we send the command.

This avoid extra stack allocations and a memcpy in the code.  The
memcpy() was causing a static analyzer warning because it was not
using the size of the destination.

type=bugfix
ticket=none
fixes=I52d2afa9f243765b63da3279496c524c77b26e20

Change-Id: Iaaeba37a842ba6abc9e53c7f018f5362964a6daf
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/222716
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
(cherry picked from commit 023605ffda133e4cea2fe29c2ec7dbcbe48988c4)
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/222778
x-iwlwifi-stack-dev: a5b2f38f100b2bb1d84696921a438c19fd3741ed