[BUGFIX] iwlwifi: pcie: correct lockdep annotation placement
When I migrated the code to use NAPI properly, I removed the same
lockdep annotations from the MSI-X interrupt handlers and instead
moved them into the NAPI processing.

This was wrong, because the sync_cmd_lockdep_map was intended to
protect us from accidentally taking the same lock while sending a
command to the firmware, and in the (threaded) interrupt handler,
so we'd deadlock since the interrupt handler couldn't tell anyone
that the command was finished.

This is all a pretty fringe scenario, but nonetheless the correct
placement of the lockdep annotations is in the interrupt handlers,
not in the NAPI poll function.

We cannot have a similar deadlock against the NAPI poll in the
first place since that is in soft-IRQ (or at least with soft-IRQs
disabled), and thus cannot take a lock that we can sleep under,
which is required for sending a sync command, so we don't need to
take any extra precautions for that case.

type=bugfix
fixes=I9c84a147288fcfb1b019572c6758f2d92949f5d7
ticket=jira:WIFI-88418
ticket=chrome-os-partner:174214212

Change-Id: I942a08407282d9a0caab911d81b429bf1c3a22f6
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/286310
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Ben Ami, Golan <golan.ben.ami@intel.com>
x-iwlwifi-stack-dev: 632ab56c07c3e5385b7742d0f9e10dd5533bb888