[BUGFIX] iwlwifi: fw: fix powersave debugfs out-of-bounds read
The code here checked only that the beginning of the field
(offsetof) fits into the data returned by the firmware, not
that the *entire* field fits. Fix this to use offsetofend()
and see that is <= the size to fix this.

type=bugfix
ticket=none
fixes=Icf5304f10e108fbcc55335f97484987c3b2bf0e2

Change-Id: If411638e985242036476192620442cbd179d9a30
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/291735
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/308813
Tested-by: Luciano Coelho <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 026214b09718aaff76edafd6985cedbe81513d57